Add Temperature to FileAttributes
Summary: .. so that appropriate implementations can return temperature
information from GetChildrenFileAttributes

Test Plan: just an API placeholder for now
