+++ include/rocksdb/types.h
@@ -110,4 +110,16 @@ enum class WriteStallCondition {
   kNormal,
 };
 
+// Temperature of a file. Used to pass to FileSystem for a different
+// placement and/or coding.
+// Reserve some numbers in the middle, in case we need to insert new tier
+// there.
+enum class Temperature : uint8_t {
+  kUnknown = 0,
+  kHot = 0x04,
+  kWarm = 0x08,
+  kCold = 0x0C,
+  kLastTemperature,
+};
+
 }  // namespace ROCKSDB_NAMESPACE
